savepoint

+++ helper_isatab.rb
@@ -3,8 +3,15 @@ module OpenTox
   # @see http://api.toxbank.net/index.php/Investigation ToxBank API Investigation
   class Application < Service
 
+
     module Helpers
       # check for investigation type
+      def subtask_uri
+        response = OpenTox::Backend::FourStore.query "SELECT ?o WHERE {<#{investigation_uri}> <#{RDF::TB}hasSubTaskURI> ?o}", "application/json"
+        result = JSON.parse(response)
+        type = result["results"]["bindings"].map {|n|  "#{n["o"]["value"]}"}
+      end
+      
       def is_isatab?
         response = OpenTox::Backend::FourStore.query "SELECT ?o WHERE {<#{investigation_uri}> <#{RDF::TB}hasInvType> ?o}", "application/json"
         result = JSON.parse(response)
@@ -46,6 +53,12 @@ module OpenTox
       end
 
       def build_gene_files
+        # delete existing json files and cancel subtask if still running
+        $logger.debug "Delete existing bio search result files."
+        subtaskuri = subtask_uri[0]
+        $logger.debug "cancel running subtask: #{subtaskuri}"
+        `curl -Lk -X PUT -d '' '#{subtaskuri}/Cancelled'` unless subtaskuri.blank?
+        jsonfiles = Dir["#{dir}/*.json"].each{|file| `rm '#{file}'`} unless jsonfiles.blank?
         $logger.debug "Start processing derived data."
         templates = get_templates "investigation"
         # locate derived data files and prepare
@@ -53,48 +66,64 @@ module OpenTox
         datafiles = Dir["#{dir}/*.txt"].each{|file| `dos2unix -k '#{file}'`}
         sparqlstring = File.read(templates["files_by_assays"]) % { :investigation_uri => investigation_uri }
         response = OpenTox::Backend::FourStore.query sparqlstring, "application/json"
-        #$logger.debug response
         datafiles = JSON.parse(response)["results"]["bindings"].map{|f| f["file"]["value"]}.uniq
         $logger.debug "datafiles: #{datafiles}"
-        #datafiles.reject{|f| f =~ /ftp\:/}
         @client = Mongo::Client.new([ '127.0.0.1:27017' ], :database => 'ToxBank', :connect => :direct)
         my = @client[params[:id]]
-        datafiles.reject!{|file| file =~ /^i_|^a_|^s_|ftp\:/}.each do |file| 
+        datafiles.reject{|file| file =~ /^i_|^a_|^s_|ftp\:/}.each do |file| 
           $logger.debug "import file: #{file}"
-          `mongoimport -d ToxBank -c #{params[:id]} --ignoreBlanks --upsert --type tsv --file '#{File.join(dir, file)}' --headerline`
+          `mongoimport -d ToxBank -c #{params[:id]} --ignoreBlanks --type tsv --file '#{File.join(dir, file)}' --headerline`
         end unless datafiles.blank?
         # building genelist
         my = @client[params[:id]]
+        genelist = []
         symbol = my.find.distinct(:Symbol)
+        symbol.each{|x| genelist << "http://onto.toxbank.net/isa/Symbol/#{x}"} unless symbol.blank?
         entrez = my.find.distinct(:Entrez)
-        genelist = symbol+entrez
-        $logger.debug genelist
+        entrez.each{|x| genelist << "http://onto.toxbank.net/isa/Entrez/#{x}"} unless entrez.blank?
+        unigene = my.find.distinct(:Unigene)
+        unigene.each{|x| genelist << "http://onto.toxbank.net/isa/Unigene/#{x}"} unless unigene.blank?
+        refseq = my.find.distinct(:RefSeq)
+        refseq.each{|x| genelist << "http://onto.toxbank.net/isa/RefSeq/#{x}"} unless refseq.blank?
+        uniprod = my.find.distinct(:Uniprod)
+        uniprod.each{|x| genelist << "http://purl.uniprod.org/uniprod/#{x}"} unless uniprod.blank?
         # write to file
-        File.open(File.join(dir, "genelist"), 'w') {|f| f.write(genelist.reject!{|g| g.to_s =~ /^NA$|^0$/}) }
-        #$logger.debug genelist
-        # TODO could be more than one assay or study
+        File.open(File.join(dir, "genelist"), 'w') {|f| f.write(genelist.flatten.compact.reject{|g| g.to_s =~ /\/NA$|\/0$/}) }
+        #TODO could be more than one assay or study
+        #TODO check files for utf-8 error
         assayfiles = Dir["#{dir}/a_*.txt"][0]
-        #file = datafiles.find{|f| f =~ /\/a_/ }
         $logger.debug assayfiles
         assay = CSV.read(assayfiles, { :col_sep => "\t", :row_sep => :auto, :headers => true, :header_converters => :symbol })
         $logger.debug assay.headers
-        #file = datafiles.find{|f| f =~ /\/s_/ }
         studyfiles = Dir["#{dir}/s_*.txt"][0]
         $logger.debug studyfiles
         study = CSV.read(studyfiles, { :col_sep => "\t", :row_sep => :auto, :headers => true, :header_converters => :symbol })
         investigationfile = Dir["#{dir}/i_*.txt"][0]
         inv = CSV.read(investigationfile, { :col_sep => "\t", :row_sep => :auto, :headers => true, :header_converters => :symbol })
         inv.find{|r| @title = r[1] if r[0] == "Investigation Title"}
-        genes = genelist
+        genes = genelist.flatten.compact.reject{|g| g.to_s =~ /\/NA$|\/0$/}
         # working with genes
-        #TODO write right gene class key e.g. Symbol: TSPAN6
         genes.each do |gene|
-          out = []
+          geneclass = (gene =~ /uniprod/ ? gene.split("/")[3] : gene.split("/")[4])
+          gene = gene.split("/").last
           unless File.exists?(File.join(dir, "#{gene}.json"))
-            a = (gene.class == String ? my.find(Symbol: "#{gene}").each{|hash| hash.delete_if{|k, v| k !~ /^p-value|^q-value|^FC/}} : my.find(Entrez: gene).each{|hash| hash.delete_if{|k, v| k !~ /^p-value|^q-value|^FC/}} )
-            unless a.blank?
+            case geneclass
+            when "Symbol"
+              a = my.find(Symbol: gene).each{|hash| hash.delete_if{|k, v| k !~ /^p-value|^q-value|^FC/}}
+            when "Uniprod"
+              a = my.find(Uniprod: gene).each{|hash| hash.delete_if{|k, v| k !~ /^p-value|^q-value|^FC/}}
+            when "Unigene"
+              a = my.find(Unigene: gene).each{|hash| hash.delete_if{|k, v| k !~ /^p-value|^q-value|^FC/}}
+            when "RefSeq"
+              a = my.find(RefSeq: gene).each{|hash| hash.delete_if{|k, v| k !~ /^p-value|^q-value|^FC/}}
+            when "Entrez"
+              a = my.find(Entrez: gene.to_i).each{|hash| hash.delete_if{|k, v| k !~ /^p-value|^q-value|^FC/}}
+            else
+              bad_request_error "Unknown gene class '#{geneclass}'"
+            end
+            unless a.to_a[0].blank?
               b = {}
-              assay[:data_transformation_name].each_with_index{|name, idx| a.to_a[0].each{|a| (b.has_key?(name) ? b[name] << [:investigation => {:type => "uri", :value => investigation_uri}, :invTitle => {:type => "literal", :value => @title}, :featureType => {:type => "uri", :value=> (("http://onto.toxbank.net/isa/pvalue" if a[0] =~ /p-value/) or ("http://onto.toxbank.net/isa/qvalue" if a[0] =~ /q-value/) or ("http://onto.toxbank.net/isa/FC" if a[0] =~ /FC/)) }, :title => {:type => "literal", :value => a[0]}, :dataTransformationName => {:type => "literal", :value => name}, :value => {:type => "literal", :value => a[1], :datatype => "http://www.w3.org/2001/XMLSchema#double"}, :gene => "Symbol:#{gene}", :sample => assay[:sample_name][idx]] : b[name] = [:investigation => {:type => "uri", :value => investigation_uri}, :invTitle => {:type => "literal", :value => @title}, :featureType => {:type => "uri", :value=> (("http://onto.toxbank.net/isa/pvalue" if a[0] =~ /p-value/) or ("http://onto.toxbank.net/isa/qvalue" if a[0] =~ /q-value/) or ("http://onto.toxbank.net/isa/FC" if a[0] =~ /FC/)) }, :title => {:type => "literal", :value => a[0]}, :dataTransformationName => {:type => "literal", :value => name}, :value => {:type => "literal", :value => a[1], :datatype => "http://www.w3.org/2001/XMLSchema#double"}, :gene => "Symbol:#{gene}", :sample => assay[:sample_name][idx]]) if a[0] =~ /\b(#{name})\b/ } }
+              assay[:data_transformation_name].each_with_index{|name, idx| a.to_a[0].each{|a| (b.has_key?(name) ? b[name] << [:investigation => {:type => "uri", :value => investigation_uri}, :invTitle => {:type => "literal", :value => @title}, :featureType => {:type => "uri", :value=> (("http://onto.toxbank.net/isa/pvalue" if a[0] =~ /p-value/) or ("http://onto.toxbank.net/isa/qvalue" if a[0] =~ /q-value/) or ("http://onto.toxbank.net/isa/FC" if a[0] =~ /FC/)) }, :title => {:type => "literal", :value => a[0]}, :dataTransformationName => {:type => "literal", :value => name}, :value => {:type => "literal", :value => a[1], :datatype => "http://www.w3.org/2001/XMLSchema#double"}, :gene => "#{geneclass}:#{gene}", :sample => assay[:sample_name][idx]] : b[name] = [:investigation => {:type => "uri", :value => investigation_uri}, :invTitle => {:type => "literal", :value => @title}, :featureType => {:type => "uri", :value=> (("http://onto.toxbank.net/isa/pvalue" if a[0] =~ /p-value/) or ("http://onto.toxbank.net/isa/qvalue" if a[0] =~ /q-value/) or ("http://onto.toxbank.net/isa/FC" if a[0] =~ /FC/)) }, :title => {:type => "literal", :value => a[0]}, :dataTransformationName => {:type => "literal", :value => name}, :value => {:type => "literal", :value => a[1], :datatype => "http://www.w3.org/2001/XMLSchema#double"}, :gene => "#{geneclass}:#{gene}", :sample => assay[:sample_name][idx]]) if a[0] =~ /\b(#{name})\b/ } }
               c = {}
               assay[:sample_name].each{|sample| study.each{|x| c[x[-1]] = {:factorValues => [{:factorname => {:type => "literal", :value => "sample TimePoint"}, :value => {:type => "literal", :value => x[28], :datatype => "http://www.w3.org/2001/XMLSchema#int"}, :unit => {:type => "literal", :value => x[29]}}, {:factorname => {:type => "literal", :value => "dose"}, :value => {:type => "literal", :value => x[23], :datatype => "http://www.w3.org/2001/XMLSchema#int"}, :unit => {:type => "literal", :value => x[24]}}, :factorname => {:type => "literal", :value => "compound"}, :value => {:type => "literal", :value => x[18]}], :cell => "#{x[2]},#{x[13]}"} if x[-1] =~ /\b(#{sample})\b/}}
               b.each{|k, v| v[0]["factorValues"] = c[v[0][:sample]][:factorValues]; v[0]["cell"] = c[v[0][:sample]][:cell]}
@@ -103,10 +132,12 @@ module OpenTox
               x = []
               b.each{|k,v| v.each{|a| x << a}}
               body = {"results" => {"bindings" => x}}
-              File.open(File.join(dir, "#{gene}.json"), 'w') {|f| f.write(JSON.pretty_generate(head.merge(body))) }
+              File.open(File.join(dir, "#{gene}.json"), 'w') {|f| f.write(JSON.pretty_generate(head.merge(body))) } if Dir.exists?(dir)
             end
           end
         end unless genes.empty?
+        $logger.debug "Delete collection #{params[:id]}."
+        my.drop
         $logger.debug "End processing derived data."
       end
 
@@ -138,6 +169,7 @@ module OpenTox
           OpenTox::Backend::FourStore.put investigation_uri, File.read(File.join(dir,nt)), "application/x-turtle"
           
           task = OpenTox::Task.run("Processing derived data",investigation_uri) do
+            $logger.debug "build_gene_files"
             build_gene_files
             # update JSON object with dashboard values
             dashboard_cache
